+++ MassEffectModder/MassEffectModder/Helpers/MemoryStream.cpp
@@ -239,7 +239,7 @@ void MemoryStream::ReadStringASCIINull(QString &str)
 void MemoryStream::ReadStringUnicode16(QString &str, qint64 count)
 {
     str = "";
-    for (int n = 0; n < count; n++)
+    for (qint64 n = 0; n < count; n++)
     {
         quint16 c = ReadUInt16();
         str += QChar(static_cast<ushort>(c));
